requirements.txt
streamlit
streamlit-option-menu
pandas
psycopg2-binary
XlsxWriter
requests
beautifulsoup4
snowflake-connector-python